Create CRTP.cpp
Defines a Derived and Comparable class. Comparable is a templated class and Derived is a CRTP class that inherits from base class Comparable with a type of Derived. 

Comparable is able to use operator< to overload all the other comparison operators (>, ==, !=, <=, and >=). As such, Derived defines its own operator< and Comparable depends on the child class to support it an operator< definition. This also caused some issues from trying to templatize Derived so as such the current implementation of Derived only supports an integer member.